duplicates handled

+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: labdello <labdello@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/10 19:18:26 by labdello          #+#    #+#             */
-/*   Updated: 2024/06/16 14:41:29 by labdello         ###   ########.fr       */
+/*   Updated: 2024/06/25 18:02:02 by labdello         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 typedef struct s_stack
 {
-	int				index;
+	size_t			index;
 	int				value;
 	struct s_stack	*next;
 	struct s_stack	*prev;
